Format type fixes from misc@mandriva.org
20070619120126-b58c9-44b9271200324439d74cb1969dc22f6f661c4de8.gz

+++ wocky/wocky-xmpp-reader.c
@@ -364,7 +364,7 @@ wocky_xmpp_reader_push(WockyXmppReader *reader,
   xmlParserCtxtPtr parser;
 
   g_assert(!priv->error);
-  DEBUG("Parsing chunk: %.*s", length, data);
+  DEBUG("Parsing chunk: %.*s", (int)length, data);
 
   parser = priv->parser;
   xmlParseChunk(parser, (const char*)data, length, FALSE);
